travel_agent_deploy/requirements.txt
google-adk[eval]
google-cloud-aiplatform[adk,agent_engines]